12_REST_API/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.29.0
Flask-Limiter==3.3.0
Flask-CORS==3.0.10
flasgger==0.9.5
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.3.1
pytest-flask==1.2.0
gunicorn==20.1.0 